Handle SQL connection issues more gracefully

+++ src/main/java/me/zford/jobs/dao/JobsDAO.java
@@ -71,25 +71,22 @@ public abstract class JobsDAO {
      * @return list of all of the names of the jobs the players are part of.
      */
     public List<JobsDAOData> getAllJobs(JobsPlayer player) {
-        ArrayList<JobsDAOData> jobs = null;
-        try{
-            JobsConnection conn = getConnection();
+        ArrayList<JobsDAOData> jobs = new ArrayList<JobsDAOData>();
+        JobsConnection conn = getConnection();
+        if (conn == null)
+            return jobs;
+        try {
             String sql = "SELECT `experience`, `level`, `job` FROM `" + prefix + "jobs` WHERE `username` = ?;";
             PreparedStatement prest = conn.prepareStatement(sql);
             prest.setString(1, player.getName());
             ResultSet res = prest.executeQuery();
-            while(res.next()){
-                if(jobs == null){
-                    jobs = new ArrayList<JobsDAOData>();
-                }
+            while (res.next()) {
                 jobs.add(new JobsDAOData(res.getString(3), res.getInt(1), res.getInt(2)));
             }
             prest.close();
             conn.close();
-        }
-        catch(SQLException e){
+        } catch (SQLException e) {
             e.printStackTrace();
-            plugin.disablePlugin();
         }
         return jobs;
     }
@@ -101,8 +98,10 @@ public abstract class JobsDAO {
      */
     public void joinJob(JobsPlayer player, Job job) {
         String sql = "INSERT INTO `" + prefix + "jobs` (`username`, `experience`, `level`, `job`) VALUES (?, ?, ?, ?);";
+        JobsConnection conn = getConnection();
+        if (conn == null)
+            return;
         try {
-            JobsConnection conn = getConnection();
             PreparedStatement prest = conn.prepareStatement(sql);
             prest.setString(1, player.getName());
             prest.setInt(2, 0);
@@ -113,7 +112,6 @@ public abstract class JobsDAO {
             conn.close();
         } catch (SQLException e) {
             e.printStackTrace();
-            plugin.disablePlugin();
         }
     }
     
@@ -123,8 +121,10 @@ public abstract class JobsDAO {
      * @param job - job that the player wishes to quit
      */
     public void quitJob(JobsPlayer player, Job job) {
-        try{
-            JobsConnection conn = getConnection();
+        JobsConnection conn = getConnection();
+        if (conn == null)
+            return;
+        try {
             String sql1 = "DELETE FROM `" + prefix + "jobs` WHERE `username` = ? AND `job` = ?;";
             PreparedStatement prest = conn.prepareStatement(sql1);
             prest.setString(1, player.getName());
@@ -132,10 +132,8 @@ public abstract class JobsDAO {
             prest.executeUpdate();
             prest.close();
             conn.close();
-        }
-        catch(SQLException e){
+        } catch(SQLException e) {
             e.printStackTrace();
-            plugin.disablePlugin();
         }       
     }
     
@@ -145,10 +143,12 @@ public abstract class JobsDAO {
      */
     public void save(JobsPlayer player) {
         String sql = "UPDATE `" + prefix + "jobs` SET `experience` = ?, `level` = ? WHERE `username` = ? AND `job` = ?;";
+        JobsConnection conn = getConnection();
+        if (conn == null)
+            return;
         try {
-            JobsConnection conn = getConnection();
             PreparedStatement prest = conn.prepareStatement(sql);
-            for(JobProgression temp: player.getJobsProgression()){
+            for (JobProgression temp: player.getJobsProgression()) {
                 prest.setInt(1, (int)temp.getExperience());
                 prest.setInt(2, temp.getLevel());
                 prest.setString(3, player.getName());
@@ -159,7 +159,6 @@ public abstract class JobsDAO {
             conn.close();
         } catch (SQLException e) {
             e.printStackTrace();
-            plugin.disablePlugin();
         }
     }
     
@@ -168,23 +167,23 @@ public abstract class JobsDAO {
      * @param job - the job
      * @return  the number of players that have a particular job
      */
-    public Integer getSlotsTaken(Job job) {
-        Integer slot = 0;
-        try{
-            JobsConnection conn = getConnection();
+    public int getSlotsTaken(Job job) {
+        int slot = 0;
+        JobsConnection conn = getConnection();
+        if (conn == null)
+            return slot;
+        try {
             String sql = "SELECT COUNT(*) FROM `" + prefix + "jobs` WHERE `job` = ?;";
             PreparedStatement prest = conn.prepareStatement(sql);
             prest.setString(1, job.getName());
             ResultSet res = prest.executeQuery();
-            if(res.next()){
+            if (res.next()) {
                 slot = res.getInt(1);
             }
             prest.close();
             conn.close();
-        }
-        catch(SQLException e){
+        } catch(SQLException e) {
             e.printStackTrace();
-            plugin.disablePlugin();
         }
         return slot;
     }
@@ -194,8 +193,13 @@ public abstract class JobsDAO {
      * @return  JobsConnection object
      * @throws SQLException 
      */
-    protected JobsConnection getConnection() throws SQLException {
-        return pool.getConnection();
+    protected JobsConnection getConnection() {
+        try {
+            return pool.getConnection();
+        } catch (SQLException e) {
+            plugin.getLogger().severe("Unable to connect to the database: "+e.getMessage());
+            return null;
+        }
     }
     
     /**
